+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 CHANGELOG
 =========
+V 1-3-5
+
+- Fixed a Null Pointer exception
+- Reload Bugs from all Platforms, even if any of them Fails
+
 V 1-3-4
 
 - OSM Notes were not correctly reloaded when an error occurred during a transaction like commenting or closing a note